+++ example/lib/src/ui/device_detail/characteristic_interaction_dialog.dart
@@ -1,0 +1,235 @@
+import 'dart:async';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_reactive_ble/flutter_reactive_ble.dart';
+import 'package:flutter_reactive_ble_example/src/ble/ble_device_interactor.dart';
+import 'package:provider/provider.dart';
+
+class CharacteristicInteractionDialog extends StatelessWidget {
+  const CharacteristicInteractionDialog({
+    required this.characteristic,
+    Key? key,
+  }) : super(key: key);
+  final QualifiedCharacteristic characteristic;
+
+  @override
+  Widget build(BuildContext context) {
+    return Consumer<BleDeviceInteractor>(builder: (context, interactor, _) {
+      return _CharacteristicInteractionDialog(
+        characteristic: characteristic,
+        readCharacteristic: interactor.readCharacteristic,
+        writeWithResponse: interactor.writeCharacterisiticWithResponse,
+        writeWithoutResponse: interactor.writeCharacterisiticWithoutResponse,
+        subscribeToCharacteristic: interactor.subScribeToCharacteristic,
+      );
+    });
+  }
+}
+
+class _CharacteristicInteractionDialog extends StatefulWidget {
+  const _CharacteristicInteractionDialog({
+    required this.characteristic,
+    required this.readCharacteristic,
+    required this.writeWithResponse,
+    required this.writeWithoutResponse,
+    required this.subscribeToCharacteristic,
+    Key? key,
+  }) : super(key: key);
+
+  final QualifiedCharacteristic characteristic;
+  final Future<List<int>> Function(QualifiedCharacteristic characteristic)
+      readCharacteristic;
+  final Future<void> Function(
+          QualifiedCharacteristic characteristic, List<int> value)
+      writeWithResponse;
+
+  final Stream<List<int>> Function(QualifiedCharacteristic characteristic)
+      subscribeToCharacteristic;
+
+  final Future<void> Function(
+          QualifiedCharacteristic characteristic, List<int> value)
+      writeWithoutResponse;
+
+  @override
+  _CharacteristicInteractionDialogState createState() =>
+      _CharacteristicInteractionDialogState();
+}
+
+class _CharacteristicInteractionDialogState
+    extends State<_CharacteristicInteractionDialog> {
+  late String readOutput;
+  late String writeOutput;
+  late String subscribeOutput;
+  late TextEditingController textEditingController;
+  late StreamSubscription<List<int>>? subscribeStream;
+
+  @override
+  void initState() {
+    readOutput = '';
+    writeOutput = '';
+    subscribeOutput = '';
+    textEditingController = TextEditingController();
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    subscribeStream?.cancel();
+    super.dispose();
+  }
+
+  Future<void> subscribeCharacteristic() async {
+    subscribeStream =
+        widget.subscribeToCharacteristic(widget.characteristic).listen((event) {
+      setState(() {
+        subscribeOutput = event.toString();
+      });
+    });
+    setState(() {
+      subscribeOutput = 'Notification set';
+    });
+  }
+
+  Future<void> readCharacteristic() async {
+    final result = await widget.readCharacteristic(widget.characteristic);
+    setState(() {
+      readOutput = result.toString();
+    });
+  }
+
+  List<int> _parseInput() {
+    return textEditingController.text
+        .split(',')
+        .map(
+          (value) => int.parse(value),
+        )
+        .toList();
+  }
+
+  Future<void> writeCharacteristicWithResponse() async {
+    await widget.writeWithResponse(widget.characteristic, _parseInput());
+    setState(() {
+      writeOutput = 'Ok';
+    });
+  }
+
+  Future<void> writeCharacteristicWithoutResponse() async {
+    await widget.writeWithoutResponse(widget.characteristic, _parseInput());
+    setState(() {
+      writeOutput = 'Done';
+    });
+  }
+
+  Widget sectionHeader(String text) => Text(
+        text,
+        style: TextStyle(fontWeight: FontWeight.bold),
+      );
+
+  List<Widget> get writeSection => [
+        sectionHeader('Write characteristic'),
+        Padding(
+          padding: const EdgeInsets.symmetric(vertical: 8.0),
+          child: TextField(
+            controller: textEditingController,
+            decoration: InputDecoration(
+              border: OutlineInputBorder(),
+              labelText: 'Value',
+            ),
+            keyboardType: TextInputType.numberWithOptions(
+              decimal: true,
+              signed: false,
+            ),
+          ),
+        ),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            ElevatedButton(
+              onPressed: writeCharacteristicWithResponse,
+              child: Text('With response'),
+            ),
+            ElevatedButton(
+              onPressed: writeCharacteristicWithoutResponse,
+              child: Text('Without response'),
+            ),
+          ],
+        ),
+        Padding(
+          padding: const EdgeInsetsDirectional.only(top: 8.0),
+          child: Text('Output: $writeOutput'),
+        ),
+      ];
+
+  List<Widget> get readSection => [
+        sectionHeader('Read characteristic'),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            ElevatedButton(
+              onPressed: readCharacteristic,
+              child: Text('Read'),
+            ),
+            Text('Output: $readOutput'),
+          ],
+        ),
+      ];
+
+  List<Widget> get subscribeSection => [
+        sectionHeader('Subscribe / notify'),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            ElevatedButton(
+              onPressed: subscribeCharacteristic,
+              child: Text('Subscribe'),
+            ),
+            Text('Output: $subscribeOutput'),
+          ],
+        ),
+      ];
+
+  Widget get divider => Padding(
+        padding: const EdgeInsets.symmetric(vertical: 12.0),
+        child: Divider(thickness: 2.0),
+      );
+
+  @override
+  Widget build(BuildContext context) {
+    return Dialog(
+      child: Padding(
+        padding: const EdgeInsets.all(20.0),
+        child: ListView(
+          shrinkWrap: true,
+          children: [
+            Text(
+              'Select an operation',
+              style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+            ),
+            Padding(
+              padding: const EdgeInsets.symmetric(vertical: 8.0),
+              child: Text(
+                widget.characteristic.characteristicId.toString(),
+              ),
+            ),
+            divider,
+            ...readSection,
+            divider,
+            ...writeSection,
+            divider,
+            ...subscribeSection,
+            divider,
+            Align(
+              alignment: Alignment.bottomRight,
+              child: Padding(
+                padding: const EdgeInsets.only(top: 20.0),
+                child: ElevatedButton(
+                    onPressed: () => Navigator.of(context).pop(),
+                    child: Text('close')),
+              ),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}