+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Main releases
+
+## 3.0.0
+
+* Breaking: Migration to null-safety.
+* Improve example app by adding discover services, read, write and subscribe to characteristic. 
+
 ## 2.7.3
 
 * Upgrade Android dependencies to comply to newer Gradle distributions.